Move Database-backed MerkleTree to kv submodule
This change moves the existing merkle tree implementation to a kv
submodule and re-exports it from the original module. This provides a
space for additional back-end implementations.
